pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
google-genai = "^1.46.0"
pydantic = "^2.12.3"
streamlit = "^1.51.0"
pytest = "^8.4.2"
python-dotenv = "^1.2.1"
matplotlib = "^3.10.7"
google-cloud-firestore = "^2.21.0"
google-cloud-datastore = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "pytest-asyncio (>=0.23.7,<1.0.0)",
    "pytest-cov (>=5.0.0,<6.0.0)"
]
